+++ src/types/Settings.ts
@@ -95,6 +95,11 @@ type Settings = {
       blockSearchInterval: number;
     };
   };
+  api: {
+    restrict: {
+      apiKey: string;
+    };
+  };
 };
 
 export default Settings;