+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.9.0] - 2022-02-02
+### Added
+- Create metrics/voters endpoint to get each voters vote count of given time interval
+- Create restrictAccess method in AuthenticationMiddleware
+- Use restrictAccess in metrics routes
+
 ## [4.8.0] - 2022-01-27
 ### Added
 - Check if balance is enough before execute replicate blocks